Create README.md
ReadMe - React Web:
We will need node.js with this application. We used links in conjunction with anchor tags(for CSS styling purposes).
We also separate the JS file tabs into their respective component folders where we deconstruct/decomposition or normalization the code by organizing the bigger file into smaller files so we can edit them easily and efficiently in the future. It is exported default to the main App.js. This process follows Model View Controller design.

The coding got changed in a Navigation point of view if we want to differentiate between simple to react website
With JavaScript, we used anchor tags to navigate while we used Links with React Web with conjuction to Switch and Router Path to navigate pages

User Guide:
Please make sure that you check the directory or files in the folder by typing “ls” in the View Terminal of Visual Studio Code. Then go inside of the direct folder by typing “cd (foldername)”.
This application just runs in a local host by typing “npm start” in the terminal.